% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendr.R
\name{create_calendar}
\alias{create_calendar}
\title{Create a calendar of non-working days}
\usage{
create_calendar(
  holidays = auckland_council_holidays(),
  weekend = c("sunday", "saturday")
)
}
\arguments{
\item{holidays}{A vector of Dates which contains the non-working days. By
default, this returns a list of non-working days for staff at Auckland
Council.}

\item{weekend}{A character vector of days that are weekends. It accepts:
\code{sunday}, \code{monday}, \code{tuesday}, \code{wednesday}, \code{thursday}, \code{friday},
\code{saturday.} By default, Saturday and Sunday are assumed to be the days of
the weekend.}
}
\description{
\code{create_calendar} creates a calendar and stores it in the
calendar register for use when calling \code{create_schedule}.
}
\examples{
# A calendar of non-working days for the first Quarter of 2023, with weekends
# set to the default (Saturdays and Sundays)
q1_2023 <- c(
  seq(as.Date("2023-01-02"), as.Date("2023-01-06"), by = "days"),
  as.Date("2023-01-30"),
  as.Date("2023-02-06")
)
create_calendar(holidays = q1_2023)
}
