% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendr-package.R
\docType{package}
\name{mendr-package}
\alias{mendr}
\alias{mendr-package}
\title{mendr: Month-End Scheduling for Business Users}
\description{
A convenient way for determining dates for month-end business tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auckland-Council-CC-Insights-Team/mendr}
  \item Report bugs at \url{https://github.com/Auckland-Council-CC-Insights-Team/mendr/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{lee.durbin@aucklandcouncil.govt.nz}

Other contributors:
\itemize{
  \item Auckland Council [copyright holder]
}

}
\keyword{internal}
